all: change accumulator implementation and fix style issues
This commit changes the implementation of Accumulator, which in
turn speeds up the search for (relatively) large files. On 1GB file
with N equal to 1000 search time dropped from ~22-25 minutes to
just ~2.5 minutes. Also, reduces the amount of memory used.

The comments are mostly in russian, so, apologies to those having
hard time reading them :D